+++ src/configuration.ts
@@ -2,7 +2,7 @@ import {
   ColorSettings,
   Sections,
   StandardSettings,
-  extSuffix,
+  extensionShortName,
   IFavoriteColors,
   favoriteColorSeparator,
   IPeacockElementAdjustments,
@@ -14,7 +14,9 @@ import {
   ISettingsIndexer,
   ElementNames,
   ColorAdjustmentOptions,
-  IElementColors
+  IElementColors,
+  ForegroundColors,
+  starterSetOfFavorites
 } from './models';
 import {
   getAdjustedColorHex,
@@ -22,7 +24,8 @@ import {
   getBackgroundHoverColorHex,
   getForegroundColorHex,
   getInactiveBackgroundColorHex,
-  getInactiveForegroundColorHex} from './color-library';
+  getInactiveForegroundColorHex
+} from './color-library';
 import * as vscode from 'vscode';
 import { Logger } from './logging';
 
@@ -63,7 +66,7 @@ export async function updateGlobalConfiguration<T>(
   value?: T | undefined
 ) {
   let config = vscode.workspace.getConfiguration();
-  const section = `${extSuffix}.${setting}`;
+  const section = `${extensionShortName}.${setting}`;
   Logger.info('Updating the user settings with the following changes:');
   Logger.info(`${section} = ${value}`, true);
   return await config.update(section, value, vscode.ConfigurationTarget.Global);
@@ -117,6 +120,22 @@ export async function updateWorkspaceConfiguration(
     );
 }
 
+export function getDarkForegroundColor() {
+  return readConfiguration<string>(StandardSettings.DarkForegroundColor, '');
+}
+
+export function getDarkForegroundColorOrOverride() {
+  return getDarkForegroundColor() || ForegroundColors.DarkForeground;
+}
+
+export function getLightForegroundColor() {
+  return readConfiguration<string>(StandardSettings.LightForegroundColor, '');
+}
+
+export function getLightForegroundColorOrOverride() {
+  return getLightForegroundColor() || ForegroundColors.LightForeground;
+}
+
 export function getKeepForegroundColor() {
   return readConfiguration<boolean>(
     StandardSettings.KeepForegroundColor,
@@ -133,7 +152,7 @@ export function getFavoriteColors() {
   let values = readConfiguration<IFavoriteColors[]>(
     StandardSettings.FavoriteColors
   );
-  const menu = values.map(pc => `${pc.name} ${sep} ${pc.value}`);
+  const menu = values.map(fav => `${fav.name} ${sep} ${fav.value}`);
   values = values || [];
   return {
     menu,
@@ -188,12 +207,40 @@ export async function updateKeepBadgeColor(value: boolean) {
 }
 
 export async function updateSurpriseMeOnStartup(value: boolean) {
-  return await updateGlobalConfiguration(StandardSettings.SurpriseMeOnStartup, value);
+  return await updateGlobalConfiguration(
+    StandardSettings.SurpriseMeOnStartup,
+    value
+  );
+}
+
+export async function updateDarkForegroundColor(value: string) {
+  return await updateGlobalConfiguration(
+    StandardSettings.DarkForegroundColor,
+    value
+  );
+}
+
+export async function updateLightForegroundColor(value: string) {
+  return await updateGlobalConfiguration(
+    StandardSettings.LightForegroundColor,
+    value
+  );
 }
 
 export async function addNewFavoriteColor(name: string, value: string) {
   const { values: favoriteColors } = getFavoriteColors();
   const newFavoriteColors = [...favoriteColors, { name, value }];
+  return await updateFavoriteColors(newFavoriteColors);
+}
+
+export async function writeRecommendedFavoriteColors(
+  overrideFavorites?: IFavoriteColors[]
+) {
+  let msg = `${extensionShortName}: Adding recommended favorite colors to user settings`;
+  Logger.info(msg);
+  vscode.window.showInformationMessage(msg);
+
+  const newFavoriteColors = removeDuplicatesToStarterSet(overrideFavorites);
   return await updateFavoriteColors(newFavoriteColors);
 }
 
@@ -242,10 +289,10 @@ export function getElementStyle(
 export function getAllSettingNames() {
   let settings = [];
   const affectedSettings = Object.values(AffectedSettings).map(
-    value => `${extSuffix}.${value}`
+    value => `${extensionShortName}.${value}`
   );
   const standardSettings = Object.values(StandardSettings).map(
-    value => `${extSuffix}.${value}`
+    value => `${extensionShortName}.${value}`
   );
   settings.push(...affectedSettings);
   settings.push(...standardSettings);
@@ -391,6 +438,51 @@ export function getExistingColorCustomizations() {
   return workspace.getConfiguration().get(Sections.workspacePeacockSection);
 }
 
+export function hasFavorites() {
+  const s = getAllUserSettings();
+  return s.favoriteColors.values.length;
+}
+
+export async function updateAffectedElements(
+  values: IPeacockAffectedElementSettings
+) {
+  await updateGlobalConfiguration(
+    AffectedSettings.ActivityBar,
+    values.activityBar
+  );
+  await updateGlobalConfiguration(AffectedSettings.StatusBar, values.statusBar);
+  await updateGlobalConfiguration(AffectedSettings.TitleBar, values.titleBar);
+
+  return true;
+}
+
+function getAllUserSettings() {
+  const favoriteColors = getFavoriteColors();
+  const elementAdjustments = getElementAdjustments();
+  const keepBadgeColor = getKeepBadgeColor();
+  const keepForegroundColor = getKeepForegroundColor();
+  const surpriseMeOnStartup = getSurpriseMeOnStartup();
+  const darkForegroundColor = getDarkForegroundColor();
+  const lightForegroundColor = getLightForegroundColor();
+  const {
+    activityBar: affectActivityBar,
+    statusBar: affectStatusBar,
+    titleBar: affectTitleBar
+  } = getAffectedElements();
+  return {
+    favoriteColors,
+    elementAdjustments,
+    keepBadgeColor,
+    keepForegroundColor,
+    surpriseMeOnStartup,
+    darkForegroundColor,
+    lightForegroundColor,
+    affectActivityBar,
+    affectStatusBar,
+    affectTitleBar
+  };
+}
+
 function getOriginalColor(color: string, adjustment: ColorAdjustment) {
   let oppositeAdjustment: ColorAdjustmentOptions;
 
@@ -408,4 +500,16 @@ function getOriginalColor(color: string, adjustment: ColorAdjustment) {
       break;
   }
   return getAdjustedColorHex(color, oppositeAdjustment);
+}
+
+function removeDuplicatesToStarterSet(overrideFavorites?: IFavoriteColors[]) {
+  let starter = overrideFavorites || starterSetOfFavorites;
+
+  const { values: existingFavoriteColors } = getFavoriteColors();
+
+  // starter set must be first, so it overrides the existing ones if their are dupes.
+  let faves = [...starter, ...existingFavoriteColors];
+  return faves.filter((fav, position, arr) => {
+    return arr.map(o => o.name).indexOf(fav.name) === position;
+  });
 }