+++ src/models/constants.ts
@@ -1,6 +1,13 @@
-export const extSuffix = 'peacock';
+export const extensionShortName = 'peacock';
+export const extensionId = 'johnpapa.vscode-peacock';
 export const favoriteColorSeparator = '->';
 
 // Matches the default inactive alpha in VS Code of 0x99
 // represented in 0-1 range for tinycolor.setAlpha()
 export const inactiveElementAlpha = 0x99 / 0xff;
+
+export const defaultAmountToLighten = 10;
+
+export const defaultSaturation = 0.5;
+
+export const peacockGreen = '#42b883';